==> recommend/EDA/01_eda.ipynb <==
import pandas as pd
import numpy as np
from sklearn.preprocessing import StandardScaler

df = pd.read_excel('data/20250408_FoodDB.xlsx')
df.head()

df_info = df[df.columns[:16]]
df_info.head()

df['업체명'].value_counts()/df.shape[0]
# 4000 개 정도가 

df_ingred = df[ ['식품명'] +  df.columns[16:].tolist()]
df_ingred.head()

df.columns[16:-8].tolist()

df['데이터구분명'].value_counts() / df.shape[0]

해당 없음 이 28프로 약 4000개의 데이터가 "해당없음"임

df['식품기원코드'].value_counts()

식품 기원 코드 -> 수집 시점 인듯 
1. 분석 (4000개)
2. 수집
3. 분석 & 수집
4. 산출
5. 산출
6. 산출
7. 산출

df['식품기원명'].value_counts()

# 질병에 따른 주의 영양소 칼럼명

diabate = [ '과당(g)', '포도당(g)', '단백질(g)', '오메가3 지방산(g)', '식이섬유(g)', '아연(mg)', 
    '당류(g)', '자당(g)', '콜레스테롤(mg)', '불포화지방(g)', '에너지(kcal)', 
    '탄수화물(g)', '트랜스지방산(g)', '마그네슘(mg)']

obsesity = [
    '지방(g)', '에리스리톨(g)', '카페인(mg)', '알룰로오스(g)', '단백질(g)', '오메가3 지방산(g)', 
    '식이섬유(g)', '당류(g)', '포화지방산(g)', '콜레스테롤(mg)', '불포화지방(g)', 
    '에너지(kcal)', '탄수화물(g)', '트랜스지방산(g)']

hypertension = [
    '비타민 C(mg)', '카페인(mg)', '나트륨(mg)', '오메가3 지방산(g)', '칼륨(mg)', 
    '식이섬유(g)', '불포화지방(g)', '콜레스테롤(mg)', '포화지방산(g)', 
    '칼슘(mg)', '트랜스지방산(g)', '마그네슘(mg)']


neg_diabate = ['과당(g)', '포도당(g)', '당류(g)', '자당(g)', '콜레스테롤(mg)', '에너지(kcal)', '탄수화물(g)', '트랜스지방산(g)']
pos_diabate = ['단백질(g)', '오메가3 지방산(g)', '식이섬유(g)', '아연(mg)', '불포화지방(g)', '마그네슘(mg)']

neg_obesity = ['지방(g)', '당류(g)', '포화지방산(g)', '콜레스테롤(mg)', '불포화지방(g)', '에너지(kcal)', '탄수화물(g)', '트랜스지방산(g)', '카페인(mg)']
pos_obesity = ['에리스리톨(g)', '알룰로오스(g)', '단백질(g)', '오메가3 지방산(g)', '식이섬유(g)']

neg_hypertension = ['카페인(mg)', '나트륨(mg)', '콜레스테롤(mg)', '포화지방산(g)', '트랜스지방산(g)']
pos_hypertension = ['비타민 C(mg)', '오메가3 지방산(g)', '칼륨(mg)', '식이섬유(g)', '불포화지방(g)', '칼슘(mg)', '마그네슘(mg)']




selected_nutrients = [
    '에너지(kcal)', '수분(g)', '단백질(g)', '지방(g)', '회분(g)', '탄수화물(g)', '당류(g)', '식이섬유(g)',
    '칼슘(mg)', '철(mg)', '인(mg)', '칼륨(mg)', '나트륨(mg)', '비타민 A(μg RAE)', '레티놀(μg)', '베타카로틴(μg)',
    '티아민(mg)', '리보플라빈(mg)', '니아신(mg)', '비타민 C(mg)', '비타민 D(μg)', '콜레스테롤(mg)', '포화지방산(g)',
    '트랜스지방산(g)', '니코틴산(mg)', '니코틴아마이드(mg)', '비오틴 / 바이오틴(μg)', '비타민 B6 (mg)', '비타민 B12(μg)',
    '엽산(μg DFE)', '콜린(mg)', '판토텐산(mg)', '비타민 D2(μg)', '비타민 D3(μg)', '비타민 E(mg α-TE)',
    '알파 토코페롤(mg)', '베타 토코페롤(mg)', '감마 토코페롤(mg)', '델타 토코페롤(mg)', '알파 토코트리에놀(mg)',
    '베타 토코트리에놀(mg)', '감마 토코트리에놀(mg)', '델타 토코트리에놀(mg)', '비타민 K(μg)', '비타민 K1(μg)',
    '비타민 K2(μg)', '갈락토오스(g)', '과당(g)', '당알콜(g)', '맥아당(g)', '알룰로오스(g)'
]

# 영양소 수치형만 선택 (식품명, 식품코드 제외)
features = df.drop(columns=['식품코드', '식품명'])

# NaN 값 0으로 대체a
features = features.fillna(0)

# 정규화
scaler = StandardScaler()
scaled_features = scaler.fit_transform(features)

# KNN 모델 학습 (n_neighbors = 11 → 자기 자신 + 유사 음식 10개)
knn = NearestNeighbors(n_neighbors=11, metric='euclidean')
knn.fit(scaled_features)

# 함수 수정: 유사 음식 10개 추출
def recommend_similar_foods(food_name, top_k=10):
    idx = df[df['식품명'] == food_name].index[0]
    distances, indices = knn.kneighbors([scaled_features[idx]])
    similar_indices = indices[0][1:top_k+1]  # 자기 자신 제외한 상위 K개
    return df.iloc[similar_indices]


# 실행 예시
recommend_similar_foods("김밥")

df[selected_nutrients]
